move example to example file test
